+++ app/src/main/AndroidManifest.xml
@@ -78,7 +78,7 @@
             </intent-filter>
         </activity>
         <activity android:name=".DraftNotesActivity" />
-        <activity android:name="net.pantasystem.milktea.user.SortedUsersActivity" />
+        <activity android:name="net.pantasystem.milktea.user.activity.SortedUsersActivity" />
         <activity android:name="net.pantasystem.milktea.antenna.AntennaEditorActivity" />
         <activity android:name="net.pantasystem.milktea.antenna.AntennaListActivity" />
         <activity android:name="net.pantasystem.milktea.setting.activities.PageSettingActivity" />
@@ -92,12 +92,12 @@
                 android:name="com.google.android.gms.oss.licenses.OssLicensesMenuActivity"
                 android:theme="@style/LicensePageTheme" />
         <activity
-                android:name="net.pantasystem.milktea.user.SearchAndSelectUserActivity"
+                android:name="net.pantasystem.milktea.user.activity.SearchAndSelectUserActivity"
                 android:windowSoftInputMode="adjustResize" />
         <activity android:name="net.pantasystem.milktea.userlist.UserListDetailActivity" />
         <activity android:name="net.pantasystem.milktea.userlist.ListListActivity" />
         <activity android:name=".FavoriteActivity" />
-        <activity android:name="net.pantasystem.milktea.user.FollowFollowerActivity" />
+        <activity android:name="net.pantasystem.milktea.user.activity.FollowFollowerActivity" />
         <activity android:name=".SearchActivity" />
         <activity
                 android:name=".SearchResultActivity"
@@ -125,7 +125,7 @@
                 android:name="net.pantasystem.milktea.setting.activities.SettingsActivity"
                 android:label="@string/title_activity_settings" />
         <activity
-                android:name="net.pantasystem.milktea.user.UserDetailActivity"
+                android:name="net.pantasystem.milktea.user.activity.UserDetailActivity"
                 android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.VIEW" />
