+++ app/src/main/java/jp/panta/misskeyandroidclient/di/module/NavigationModule.kt
@@ -5,9 +5,9 @@ import dagger.Module
 import dagger.hilt.InstallIn
 import dagger.hilt.android.components.ActivityComponent
 import jp.panta.misskeyandroidclient.MainNavigationImpl
-import net.pantasystem.milktea.user.SearchAndSelectUserNavigationImpl
+import net.pantasystem.milktea.user.activity.SearchAndSelectUserNavigationImpl
 import jp.panta.misskeyandroidclient.SearchNavigationImpl
-import net.pantasystem.milktea.user.UserDetailNavigationImpl
+import net.pantasystem.milktea.user.activity.UserDetailNavigationImpl
 import net.pantasystem.milktea.common_navigation.MainNavigation
 import net.pantasystem.milktea.common_navigation.SearchAndSelectUserNavigation
 import net.pantasystem.milktea.common_navigation.SearchNavigation
