+++ modules/features/user/src/main/java/net/pantasystem/milktea/user/FollowFollowerFragment.kt
@@ -19,6 +19,8 @@ import net.pantasystem.milktea.app_store.user.RequestType
 import net.pantasystem.milktea.app_store.user.from
 import net.pantasystem.milktea.app_store.user.string
 import net.pantasystem.milktea.model.user.User
+import net.pantasystem.milktea.user.compose.UserDetailCardPageableList
+import net.pantasystem.milktea.user.compose.UserDetailCardPageableListAction
 import net.pantasystem.milktea.user.viewmodel.FollowFollowerViewModel
 import net.pantasystem.milktea.user.viewmodel.provideFactory
 import javax.inject.Inject
