--- modules/features/user/src/main/java/net/pantasystem/milktea/user/ShowUserDetails.kt
@@ -1,7 +1,0 @@
-package net.pantasystem.milktea.user
-import net.pantasystem.milktea.model.user.User
-
-interface ShowUserDetails {
-
-    fun show(userId: User.Id?)
-}
