+++ modules/features/user/src/main/java/net/pantasystem/milktea/user/SortedUsersFragment.kt
@@ -15,6 +15,8 @@ import net.pantasystem.milktea.common.ResultState
 import net.pantasystem.milktea.common.StateContent
 import net.pantasystem.milktea.model.user.query.FindUsersQuery
 import net.pantasystem.milktea.model.user.query.from
+import net.pantasystem.milktea.user.compose.UserDetailCardList
+import net.pantasystem.milktea.user.compose.UserDetailCardListAction
 import net.pantasystem.milktea.user.viewmodel.SortedUsersViewModel
 import net.pantasystem.milktea.user.viewmodel.providerViewModel
 import javax.inject.Inject
