+++ modules/features/user/src/main/java/net/pantasystem/milktea/user/compose/user_detail_card_list.kt
@@ -1,13 +1,14 @@
-package net.pantasystem.milktea.user
+package net.pantasystem.milktea.user.compose
 
-import androidx.compose.foundation.layout.*
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.rememberLazyListState
 import androidx.compose.material.CircularProgressIndicator
 import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.runtime.snapshotFlow
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.ExperimentalComposeUiApi
 import androidx.compose.ui.Modifier
@@ -15,47 +16,34 @@ import androidx.compose.ui.input.nestedscroll.nestedScroll
 import androidx.compose.ui.platform.rememberNestedScrollInteropConnection
 import com.google.accompanist.swiperefresh.SwipeRefresh
 import com.google.accompanist.swiperefresh.rememberSwipeRefreshState
-import kotlinx.coroutines.flow.distinctUntilChanged
-import kotlinx.coroutines.flow.launchIn
-import kotlinx.coroutines.flow.onEach
-import net.pantasystem.milktea.common.PageableState
+import net.pantasystem.milktea.common.ResultState
 import net.pantasystem.milktea.common.StateContent
-import net.pantasystem.milktea.common.ui.isScrolledToTheEnd
 import net.pantasystem.milktea.model.user.User
 
-sealed interface UserDetailCardPageableListAction {
+sealed interface UserDetailCardListAction {
     data class CardAction(
         val cardAction: UserDetailCardAction
-    ) : UserDetailCardPageableListAction
+    ) : UserDetailCardListAction
 
-    object OnBottomReached : UserDetailCardPageableListAction
-    object Refresh : UserDetailCardPageableListAction
+    object Refresh : UserDetailCardListAction
 }
 
 @OptIn(ExperimentalComposeUiApi::class)
 @Composable
-fun UserDetailCardPageableList(
-    pageableState: PageableState<List<User.Id>>,
+fun UserDetailCardList(
+    pageableState: ResultState<List<User.Id>>,
     users: List<User.Detail>,
     isUserNameMain: Boolean,
-    onAction: (UserDetailCardPageableListAction) -> Unit,
+    onAction: (UserDetailCardListAction) -> Unit,
 ) {
     val scrollController = rememberLazyListState()
-    LaunchedEffect(key1 = null) {
-        snapshotFlow {
-            scrollController.isScrolledToTheEnd()
-        }.distinctUntilChanged().onEach {
-            if (it) {
-                onAction(UserDetailCardPageableListAction.OnBottomReached)
-            }
-        }.launchIn(this)
-    }
+
 
     when (pageableState.content) {
         is StateContent.Exist -> {
             SwipeRefresh(
-                state = rememberSwipeRefreshState(isRefreshing = pageableState is PageableState.Loading.Init),
-                onRefresh = { onAction(UserDetailCardPageableListAction.Refresh) },
+                state = rememberSwipeRefreshState(isRefreshing = pageableState is ResultState.Loading),
+                onRefresh = { onAction(UserDetailCardListAction.Refresh) },
                 modifier = Modifier
                     .nestedScroll(rememberNestedScrollInteropConnection())
                     .fillMaxSize()
@@ -66,17 +54,9 @@ fun UserDetailCardPageableList(
                             userDetail = users[i],
                             isUserNameMain = isUserNameMain,
                             onAction = {
-                                onAction(UserDetailCardPageableListAction.CardAction(it))
+                                onAction(UserDetailCardListAction.CardAction(it))
                             },
                         )
-                    }
-                    item {
-                        Row(
-                            Modifier.fillMaxWidth(),
-                            horizontalArrangement = Arrangement.Center,
-                        ) {
-                            CircularProgressIndicator()
-                        }
                     }
                 }
             }
@@ -88,14 +68,14 @@ fun UserDetailCardPageableList(
                 horizontalAlignment = Alignment.CenterHorizontally
             ) {
                 when (pageableState) {
-                    is PageableState.Error -> {
+                    is ResultState.Error -> {
                         Text("Error")
                         Text(pageableState.throwable.toString())
                     }
-                    is PageableState.Fixed -> {
+                    is ResultState.Fixed -> {
                         Text("Content is empty")
                     }
-                    is PageableState.Loading -> {
+                    is ResultState.Loading -> {
                         CircularProgressIndicator()
                     }
                 }
