+++ modules/features/user/src/main/java/net/pantasystem/milktea/user/compose/user_detail_card_pageable_list.kt
@@ -1,14 +1,13 @@
-package net.pantasystem.milktea.user
+package net.pantasystem.milktea.user.compose
 
-
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.rememberLazyListState
 import androidx.compose.material.CircularProgressIndicator
 import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.runtime.snapshotFlow
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.ExperimentalComposeUiApi
 import androidx.compose.ui.Modifier
@@ -16,34 +15,47 @@ import androidx.compose.ui.input.nestedscroll.nestedScroll
 import androidx.compose.ui.platform.rememberNestedScrollInteropConnection
 import com.google.accompanist.swiperefresh.SwipeRefresh
 import com.google.accompanist.swiperefresh.rememberSwipeRefreshState
-import net.pantasystem.milktea.common.ResultState
+import kotlinx.coroutines.flow.distinctUntilChanged
+import kotlinx.coroutines.flow.launchIn
+import kotlinx.coroutines.flow.onEach
+import net.pantasystem.milktea.common.PageableState
 import net.pantasystem.milktea.common.StateContent
+import net.pantasystem.milktea.common.ui.isScrolledToTheEnd
 import net.pantasystem.milktea.model.user.User
 
-sealed interface UserDetailCardListAction {
+sealed interface UserDetailCardPageableListAction {
     data class CardAction(
         val cardAction: UserDetailCardAction
-    ) : UserDetailCardListAction
+    ) : UserDetailCardPageableListAction
 
-    object Refresh : UserDetailCardListAction
+    object OnBottomReached : UserDetailCardPageableListAction
+    object Refresh : UserDetailCardPageableListAction
 }
 
 @OptIn(ExperimentalComposeUiApi::class)
 @Composable
-fun UserDetailCardList(
-    pageableState: ResultState<List<User.Id>>,
+fun UserDetailCardPageableList(
+    pageableState: PageableState<List<User.Id>>,
     users: List<User.Detail>,
     isUserNameMain: Boolean,
-    onAction: (UserDetailCardListAction) -> Unit,
+    onAction: (UserDetailCardPageableListAction) -> Unit,
 ) {
     val scrollController = rememberLazyListState()
-
+    LaunchedEffect(key1 = null) {
+        snapshotFlow {
+            scrollController.isScrolledToTheEnd()
+        }.distinctUntilChanged().onEach {
+            if (it) {
+                onAction(UserDetailCardPageableListAction.OnBottomReached)
+            }
+        }.launchIn(this)
+    }
 
     when (pageableState.content) {
         is StateContent.Exist -> {
             SwipeRefresh(
-                state = rememberSwipeRefreshState(isRefreshing = pageableState is ResultState.Loading),
-                onRefresh = { onAction(UserDetailCardListAction.Refresh) },
+                state = rememberSwipeRefreshState(isRefreshing = pageableState is PageableState.Loading.Init),
+                onRefresh = { onAction(UserDetailCardPageableListAction.Refresh) },
                 modifier = Modifier
                     .nestedScroll(rememberNestedScrollInteropConnection())
                     .fillMaxSize()
@@ -54,9 +66,17 @@ fun UserDetailCardList(
                             userDetail = users[i],
                             isUserNameMain = isUserNameMain,
                             onAction = {
-                                onAction(UserDetailCardListAction.CardAction(it))
+                                onAction(UserDetailCardPageableListAction.CardAction(it))
                             },
                         )
+                    }
+                    item {
+                        Row(
+                            Modifier.fillMaxWidth(),
+                            horizontalArrangement = Arrangement.Center,
+                        ) {
+                            CircularProgressIndicator()
+                        }
                     }
                 }
             }
@@ -68,14 +88,14 @@ fun UserDetailCardList(
                 horizontalAlignment = Alignment.CenterHorizontally
             ) {
                 when (pageableState) {
-                    is ResultState.Error -> {
+                    is PageableState.Error -> {
                         Text("Error")
                         Text(pageableState.throwable.toString())
                     }
-                    is ResultState.Fixed -> {
+                    is PageableState.Fixed -> {
                         Text("Content is empty")
                     }
-                    is ResultState.Loading -> {
+                    is PageableState.Loading -> {
                         CircularProgressIndicator()
                     }
                 }
