+++ modules/features/user/src/main/res/layout/activity_follow_follower.xml
@@ -5,7 +5,7 @@
             xmlns:tools="http://schemas.android.com/tools"
             android:layout_width="match_parent"
             android:layout_height="match_parent"
-            tools:context="net.pantasystem.milktea.user.FollowFollowerActivity"
+            tools:context="net.pantasystem.milktea.user.activity.FollowFollowerActivity"
             android:id="@+id/layoutBase">
 
         <androidx.appcompat.widget.Toolbar
