+++ modules/features/user/src/main/res/layout/activity_sorted_uses.xml
@@ -5,7 +5,7 @@
             xmlns:tools="http://schemas.android.com/tools"
             android:layout_width="match_parent"
             android:layout_height="match_parent"
-            tools:context="net.pantasystem.milktea.user.SortedUsersActivity">
+            tools:context="net.pantasystem.milktea.user.activity.SortedUsersActivity">
         <com.google.android.material.appbar.AppBarLayout
                 android:id="@+id/sortedUsersAppBar"
                 android:layout_width="match_parent"
