+++ modules/features/user/src/main/res/layout/activity_user_detail.xml
@@ -13,7 +13,7 @@
     <androidx.coordinatorlayout.widget.CoordinatorLayout
             android:layout_width="match_parent"
             android:layout_height="match_parent"
-            tools:context="net.pantasystem.milktea.user.UserDetailActivity"
+            tools:context="net.pantasystem.milktea.user.activity.UserDetailActivity"
             >
 
         <com.google.android.material.appbar.AppBarLayout
